+++ src/Maker/MakeForm.php
@@ -11,8 +11,10 @@
 
 namespace Symfony\Bundle\MakerBundle\Maker;
 
+use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Symfony\Bundle\MakerBundle\ConsoleStyle;
 use Symfony\Bundle\MakerBundle\DependencyBuilder;
+use Symfony\Bundle\MakerBundle\Doctrine\DoctrineEntityHelper;
 use Symfony\Bundle\MakerBundle\Generator;
 use Symfony\Bundle\MakerBundle\InputConfiguration;
 use Symfony\Bundle\MakerBundle\Str;
@@ -20,6 +22,7 @@ use Symfony\Bundle\MakerBundle\Validator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Validator\Validation;
 
@@ -29,6 +32,13 @@ use Symfony\Component\Validator\Validation;
  */
 final class MakeForm extends AbstractMaker
 {
+    private $entityHelper;
+
+    public function __construct(DoctrineEntityHelper $entityHelper)
+    {
+        $this->entityHelper = $entityHelper;
+    }
+
     public static function getCommandName(): string
     {
         return 'make:form';
@@ -39,8 +49,27 @@ final class MakeForm extends AbstractMaker
         $command
             ->setDescription('Creates a new form class')
             ->addArgument('name', InputArgument::OPTIONAL, sprintf('The name of the form class (e.g. <fg=yellow>%sType</>)', Str::asClassName(Str::getRandomTerm())))
+            ->addArgument('bound-class', InputArgument::OPTIONAL, 'The name of Entity or custom model class that the new form will be bound to (empty for none)')
             ->setHelp(file_get_contents(__DIR__.'/../Resources/help/MakeForm.txt'))
         ;
+
+        $inputConf->setArgumentAsNonInteractive('bound-class');
+    }
+
+    public function interact(InputInterface $input, ConsoleStyle $io, Command $command)
+    {
+        if (null === $input->getArgument('bound-class')) {
+            $argument = $command->getDefinition()->getArgument('bound-class');
+
+            $entities = $this->entityHelper->getEntitiesForAutocomplete();
+
+            $question = new Question($argument->getDescription());
+            $question->setValidator(function ($answer) use ($entities) {return Validator::existsOrNull($answer, $entities); });
+            $question->setAutocompleterValues($entities);
+            $question->setMaxAttempts(3);
+
+            $input->setArgument('bound-class', $io->askQuestion($question));
+        }
     }
 
     public function generate(InputInterface $input, ConsoleStyle $io, Generator $generator)
@@ -51,19 +80,33 @@ final class MakeForm extends AbstractMaker
             'Type'
         );
 
-        $entityClassNameDetails = $generator->createClassNameDetails(
-            $formClassNameDetails->getRelativeNameWithoutSuffix(),
-            'Entity\\'
-        );
+        $formFields = ['field_name'];
+        $boundClassVars = [];
+
+        $boundClass = $input->getArgument('bound-class');
+
+        if (null !== $boundClass) {
+            $boundClassDetails = $generator->createClassNameDetails(
+                $boundClass,
+                'Entity\\'
+            );
+
+            $doctrineEntityDetails = $this->entityHelper->createDoctrineDetails($boundClassDetails->getFullName());
+
+            if (null !== $doctrineEntityDetails) {
+                $formFields = $doctrineEntityDetails->getFormFields();
+            }
+
+            $boundClassVars = [
+                'bounded_full_class_name' => $boundClassDetails->getFullName(),
+                'bounded_class_name' => $boundClassDetails->getShortName(),
+            ];
+        }
 
         $generator->generateClass(
             $formClassNameDetails->getFullName(),
             'form/Type.tpl.php',
-            [
-                'entity_class_exists' => class_exists($entityClassNameDetails->getFullName()),
-                'entity_full_class_name' => $entityClassNameDetails->getFullName(),
-                'entity_class_name' => $entityClassNameDetails->getShortName(),
-            ]
+            array_merge(['form_fields' => $formFields], $boundClassVars)
         );
 
         $generator->writeChanges();
@@ -88,6 +131,12 @@ final class MakeForm extends AbstractMaker
             Validation::class,
             'validator',
             // add as an optional dependency: the user *probably* wants validation
+            false
+        );
+
+        $dependencies->addClassDependency(
+            DoctrineBundle::class,
+            'orm',
             false
         );
     }